frontend/requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=0.24.0
tensorflow>=2.8.0
torch>=1.9.0
transformers>=4.11.0

# System monitoring
psutil>=5.8.0
py-cpuinfo>=8.0.0
GPUtil>=1.4.0
netifaces>=0.11.0
speedtest-cli>=2.1.3

# Security
cryptography>=3.4.0
pyOpenSSL>=20.0.0
requests>=2.26.0
beautifulsoup4>=4.9.0

# Data processing
pyyaml>=5.4.0
python-dotenv>=0.19.0
tqdm>=4.62.0
colorama>=0.4.4

# Reporting
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.3.0
jinja2>=3.0.0
weasyprint>=52.0

# Testing
pytest>=6.2.0
pytest-cov>=2.12.0
pytest-mock>=3.6.0

# Development
black>=21.7b0
flake8>=3.9.0
mypy>=0.910
isort>=5.9.0
pre-commit>=2.15.0

# Logging and monitoring
loguru>=0.5.0
prometheus-client>=0.11.0
statsd>=3.3.0

# API and web
fastapi>=0.68.0
uvicorn>=0.15.0
websockets>=9.1.0
aiohttp>=3.7.0

# Database
sqlalchemy>=1.4.0
alembic>=1.7.0
asyncpg>=0.24.0

# Utilities
python-dateutil>=2.8.0
pytz>=2021.1
tqdm>=4.62.0
rich>=10.0.0